clearInputField when adding movies to the lists, transformed ThemeEditor to typescript, and fixed ToolTip for change moviegroup name button
